Synchronize the client and server after the send/recv test
There exists a race condition in the donard rdma client/server protocol:

The client will immediately start issuing RDMA WRITEs after it sends the
COMMON_A2/B2 SEND message.  Further, it uses the same server buffer for
the target/sink of the RDMA WRITEs as the server used for the A2/B2 RECV.
So the server occasionally gets the RDMA WRITE data in its buffer _before_
it validates the A2/B2 data in common_test_recv().

The fix is to sync the client and server after the server validates
the A2/B2 and only then start the RDMA tests.  This is done with a 0B
send/recv "go ahead" exchange.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>